requirements.txt
PyYAML==5.3.1
mysql-connector-python==8.0.20